float support# Please enter the commit message for your changes. Lines starting

+++ utils/losses.py
@@ -91,7 +91,7 @@ def combined_surface_gDice_loss_with_mauer(y_true, y_pred):
 
 def dice_coef(y_true, y_pred):
     smooth = 1e-8
-    y_true_f = K.flatten(y_true)
+    y_true_f = K.flatten(K.cast(y_true, dtype='float32'))
     y_pred_f = K.flatten(y_pred)
     intersection = K.sum(y_true_f * y_pred_f)
     return (2. * intersection + smooth) / (K.sum(y_true_f) + K.sum(y_pred_f) + smooth)
